% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{anchor}
\alias{anchor}
\title{Create an anchor object}
\usage{
anchor(model = "default", depth = 0)
}
\arguments{
\item{model}{character value indicating the model of the anchor.
At present, only \code{"default"} is permitted, and this is just
based on a guess on height (when laid on its side).}

\item{depth}{numeric value giving water depth in m.}
}
\value{
an object of the \code{"mooring"} class, with \code{type} equal to \code{"anchor"}.
}
\description{
This must be the first element of a mooring constructed with
\code{\link[=mooring]{mooring()}}.
}
\seealso{
Other functions that create mooring objects: 
\code{\link{chain}()},
\code{\link{float}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
