% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor.R
\name{anchorWeight}
\alias{anchorWeight}
\title{Find Anchor Weight of a Mooring}
\usage{
anchorWeight(m)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}
}
\value{
\code{anchorWeight} returns the weight of the mooring anchor,
in kg.
}
\description{
Find Anchor Weight of a Mooring
}
\author{
Dan Kelley
}
