% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.R
\name{angle}
\alias{angle}
\title{Angle of mooring elements}
\usage{
angle(m, stagnant = FALSE, skipWire = FALSE)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{stagnant}{logical value indicating whether to return
values for the mooring configuration in the stagnant case
(with zero velocity) or the present case (possibly with
nonzero velocity).}

\item{skipWire}{logical value indicating whether to skip wire elements, i.e. to return values just for floats, instruments, etc.}
}
\value{
a numeric vector of the angle in radians.
}
\description{
This is the angle (in radians) between the element and
the vertical.
}
\examples{
library(mooring)
m <- mooring(anchor(), wire(length = 100), float("HMB 20"), waterDepth = 120) |> knockdown(u = 1)
angle(m)

}
\author{
Dan Kelley
}
