% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Run a GUI app for interactive simulations}
\usage{
app()
}
\description{
This makes a simple mooring with an anchor, a line, and a float, in
a depth-uniform current, or one of a few other current profiles.
Sliders are provided for the adjustment of line length and current speed,
and pulldown menus are provided to adjust wire and float types.
}
\author{
Dan Kelley
}
