% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app1.R
\name{app1}
\alias{app1}
\title{Run a GUI app for interactive simulations}
\usage{
app1()
}
\description{
This makes a simple mooring with an anchor, a line, and a float, in
a variety of current profiles. It differs from \code{\link[=app2]{app2()}} and \code{\link[=app2bs]{app2bs()}}
in having an instrument as well as a float.
}
\seealso{
Other interactive apps: 
\code{\link{app2}()},
\code{\link{app2bs}()}
}
\author{
Dan Kelley
}
\concept{interactive apps}
