% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app2bs.R
\name{app2bs}
\alias{app2bs}
\title{Run a two-element GUI app for interactive simulations}
\usage{
app2bs(debug = FALSE)
}
\arguments{
\item{debug}{logical value indicating whether to print debugging
output.  This can be toggled with the keystroke 'd'.}
}
\description{
This makes a simple mooring with an anchor, an instrument, and
a float, in a variety of current profiles. It differs from \code{\link[=app1]{app1()}}
in having the instrument, and in having some interface changes.
}
\author{
Dan Kelley
}
