% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Horizontally-projected area of mooring elements}
\usage{
area(m, phi = 0)
}
\arguments{
\item{m}{either a mooring object, created by \code{\link[=mooring]{mooring()}}, or a
mooring-element object, created with}

\item{phi}{angle (in degrees) of element compared with a vertical line. For
is used for wire, chain, and intrument elements, but not for floats, which
are assumed to be spherical.}
}
\value{
\code{area} returns a numeric value of the side-view area, in m^2.
}
\description{
The areas are not computed by this function, but rather looked up
for each element.  To learn how areas are computed during
setup, see the help page for
\code{\link[=anchor]{anchor()}},
\code{\link[=chain]{chain()}},
\code{\link[=connector]{connector()}},
\code{\link[=float]{float()}},
\code{\link[=instrument]{instrument()}},
\code{\link[=misc]{misc()}}, or
\code{\link[=release]{release()}}.
}
\details{
For a summary of characteristics of the predefined models,
see the vignette named "Default Values for Mooring Elements".
}
\examples{
library(mooring)
area(float())


}
\author{
Dan Kelley
}
