% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{buoyancy}
\alias{buoyancy}
\title{Return buoyancy of elements in mooring, in kg.}
\usage{
buoyancy(m)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class.}
}
\value{
a numeric vector of buoyancy, in kg.
}
\description{
This is a direct lookup for all types except \code{"wire"}.
In the wire case, the stored buoyancy is in kg/m and therefore
the length of the wire is multiplied by that value.
}
\examples{
library(mooring)
m <- anchor(depth=120) + wire(length=100) + float("HMB 20")
buoyancy(m)

}
\author{
Dan Kelley
}
