% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{buoyancyCalculation}
\alias{buoyancyCalculation}
\title{Compute buoyancy of object given in-air properties}
\usage{
buoyancyCalculation(weight = 16.3, rho = 2400, rhoWater = 1027)
}
\arguments{
\item{weight}{numeric value giving the object's in-air weight, kg.}

\item{rho}{object's density in kg/m^3. The default is a value for steel.}

\item{rhoWater}{water density in kg/m^3. The default is a rough value
for seawater.}
}
\value{
\code{buoyancyCalculation} returns the in-seawater buoyancy of the object, in kg.
}
\description{
Compute buoyancy of object given in-air properties
}
\examples{
# 8x8x16 inch cinder block weighs 16.3kg
in2m <- 0.0254
buoyancyCalculation(16.3, 2400) # -9.32 kg
# 8x8x16 inch concrete block weighs 23.1kg
buoyancyCalculation(23.1, 2400) # -13.2 kg

}
\author{
Dan Kelley
}
