% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{cable}
\alias{cable}
\title{Create a cable object}
\usage{
cable(length = NULL, model = "Mooring Systems 3X19 3/16")
}
\arguments{
\item{length}{numeric value indicating the length of cable, in metres.}

\item{model}{character value indicating the model of the cable.}
}
\value{
an object of the \code{"mooring"} class, with \code{type} equal to \code{"cable"}.
}
\description{
Create a cable object
}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{float}()},
\code{\link{release}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
