% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{CD}
\alias{CD}
\title{Get mooring/element drag coefficient}
\usage{
CD(m)
}
\arguments{
\item{m}{either a mooring object, created by \code{\link[=mooring]{mooring()}}, or a
mooring-element object, created with \code{\link[=anchor]{anchor()}}, \code{\link[=chain]{chain()}}, \code{\link[=release]{release()}},
\code{\link[=wire]{wire()}}, \code{\link[=instrument]{instrument()}}, or \code{\link[=float]{float()}}.}
}
\value{
\code{CD} returns a numeric vector of drag coefficient(s).
}
\description{
Get mooring/element drag coefficient
}
\examples{
library(mooring)
CD(float())

}
\author{
Dan Kelley
}
