% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamped.R
\name{clamped}
\alias{clamped}
\title{Clamp an Instrument on a Wire}
\usage{
clamped(instrument)
}
\arguments{
\item{instrument}{a \code{mooring} object of subtype \code{instrument},
as created with \code{\link[=instrument]{instrument()}}.}
}
\value{
\code{clamped} returns a copy of its input, but with
the \code{height} value set to 0.
}
\description{
This changes the \code{height} field of an \code{instrument} object to 0,
so that it does not contribute to the overall length of the mooring.
Using this in deep water is unlikely to make much difference,
but it is handy for shallow-water cases, if the device was
clamped on to the wire, but the \code{height} field of the
\code{instrument} object used to represent it has non-zero \code{height}.
The example illustrates this, for mooring in very shallow water.
}
\examples{
library(mooring)
a <- anchor("my anchor", buoyancy = -50, height = 0.1, CD = 0)
w <- function(length) wire(length = length)
i <- instrument("SBE37 microcat clamp-on style")
f <- float("my float", buoyancy = 20, height = 0.2, area = 0.2^2, CD = 1.3)
# Construct unclamped mooring (m) and clamped mooring (M).
mooring(a, w(20), i, w(20), f, waterDepth = 50)
mooring(a, w(20), clamped(i), w(20), f, waterDepth = 50)

}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
