% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{connector}
\alias{connector}
\title{Create a connector object}
\usage{
connector(
  model = "swivel",
  buoyancy = NULL,
  height = NULL,
  width = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the type of connector.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If this is \code{"?X"}, where X is a set of
characters, then \code{\link[=findElement]{findElement()}} is called to do a fuzzy search,
with the \code{search} argument set to \code{"connector"}.
(3) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(4) Otherwise, a new connector object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancy}{numeric value indicating buoyancy.  For
wire elements, this is buoyancy per length, expressed in kg/m.
For other elements, it is buoyancy, expressed in kg.
Note that \code{buoyancy} is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height (in m) of the element.
This value is used in the computation of (a) the mooring geometry
and (b) the area of the element (which is in turn used in
the drag computation).
Note that \code{height} is ignored if \code{model} is recognized.}

\item{width}{numeric value for the width (in m)
of the element as viewed from above.
This is multiplied by \code{height} to compute the
area of the element, as part of the drag computation.
Note that \code{width} is ignored if \code{model} is recognized.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}
}
\value{
\code{connector} returns an object of the \code{"mooring"} class and \code{"connector"} subclass.
}
\description{
Create an object that describes mooring connector such as shackles,
either by looking up a known object from the database, or defining a new type.
}
\details{
Many of the built-in items have \code{height} and \code{width} defined
to be zero, because the sources did not list these things.  Likely, this is not
terribly important, because these values are only used in computing drag, and
connectors tend to be small, and are not likely to have significant drag compared
with tens or hundreds of meters of wire, or with large floats.

Also, it is worth noting that there are built-in connector objects that might not
be thought of as connectors, e.g. \code{"ballast weight"}.
}
\section{Data sources}{


To learn more about the data used for the built-in
connector
objects, use e.g.
\preformatted{
data(mooringElements)
mooringElements$connectors
}
}

\examples{
library(mooring)
# List known chain types
connector("?")

}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{float}()},
\code{\link{instrument}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
