% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{depth}
\alias{depth}
\title{Return depth, in m, of elements in mooring.}
\usage{
depth(m, stagnant = FALSE)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{stagnant}{logical value indicating whether to return
values for the mooring configuration in the stagnant case
(with zero velocity) or the present case (possibly with
nonzero velocity).}
}
\value{
a numeric vector of depth in metres.
}
\description{
This is the depth of the \emph{top} of the element. See also
\code{\link[=z]{z()}}, which is the negative of the result from \code{depth()}.
}
\examples{
library(mooring)
m <- mooring(anchor(depth=120), wire(length=100), float("HMB 20"))
depth(m)

}
\author{
Dan Kelley
}
