% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretise.R
\name{discretise}
\alias{discretise}
\alias{discretize}
\title{Discretise chain and wire portions of a mooring}
\usage{
discretise(m, by = 1)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{by}{numeric value giving the rough size of the chunks,
which is 1m by default. This is considered in the context of
the total length of the element, \code{L} say. If \code{L/by} exceeds
20, then chunks of length \code{by} are used.  Otherwise,
20 chunks, each of length \code{L/20}, are used.}
}
\value{
an object of the \code{"mooring"} class, identical
to \code{m} except that wire portions are chopped up into shorter
pieces.
}
\description{
Break up \code{chain} and \code{wire} portions of a mooring into smaller chunks,
so that the deformation by a current can be traced more
accurately by \code{\link[=knockdown]{knockdown()}}.
}
\author{
Dan Kelley
}
