% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{drag}
\alias{drag}
\title{Drag on mooring elements}
\usage{
drag(m, u, rho = 1027)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{u}{either a non-negative number indicating depth-independent velocity,
or a function that returns that velocity as a function of depth (m)
below the ocean surface.}

\item{rho}{a numeric value giving
water density in kg/m^3. The default value of 1027 is
likely to be sufficient for most calculations, given that
oceanographic variations of density are typically well under 1
percent, although it might make sense to use a smaller value
in lake applications.  Note that the overall force balance
is more sensitive to other things, as discussed in the \dQuote{Details}
section of the documentation for \code{\link[=buoyancy]{buoyancy()}}.}
}
\description{
This looks up element areas with \code{\link[=area]{area()}} and then computes
\eqn{(1/2)*area*rho*CD*u^2}{(1/2)*area*rho*CD*u^2}
}
\author{
Dan Kelley
}
