% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{instrument}
\alias{instrument}
\title{Create an instrument object}
\usage{
instrument(
  model = "sbe37 microcat clamp-on style",
  buoyancy = NULL,
  height = NULL,
  area = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the type of instrument.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If this is \code{"?X"}, where X is a set of
characters, then \code{\link[=findElement]{findElement()}} is called to do a fuzzy search,
with the \code{search} argument set to \code{"instrument"}.
(3) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(4) Otherwise, a new instrument object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancy}{numeric value indicating buoyancy.  For
wire elements, this is buoyancy per length, expressed in kg/m.
For other elements, it is buoyancy, expressed in kg.
Note that \code{buoyancy} is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height (in m) of the element.
This value is used in the computation of (a) the mooring geometry
and (b) the area of the element (which is in turn used in
the drag computation).
Note that \code{height} is ignored if \code{model} is recognized.}

\item{area}{a numerical value indicating the horizontal instrument area
in \eqn{m^2}{m^2}.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}
}
\value{
\code{instrument} returns an object of the \code{"mooring"} class and \code{"instrument"} subclass.
}
\description{
Create an instrument object,
either by looking up a known object from the database, or defining a new type.
}
\section{Data sources}{


To learn more about the data used for the built-in
instrument
objects, use e.g.
\preformatted{
data(mooringElements)
mooringElements$instruments
}
}

\examples{
library(mooring)
# List known instrument types
instrument("?")

}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{connector}()},
\code{\link{float}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
