% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.anchor}
\alias{is.anchor}
\title{Determine which Mooring Elements are Anchors}
\usage{
is.anchor(m)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}
}
\value{
a logical vector of the same length as \code{m}, indicating
whether each element is of the \code{anchor} sub-class.
}
\description{
Determine which Mooring Elements are Anchors
}
\examples{
library(mooring)
m <- mooring(anchor(), wire(length = 100), float("HMB 20"), waterDepth = 120)
is.anchor(m)

}
\author{
Dan Kelley
}
