% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.float}
\alias{is.float}
\title{Determine which Mooring Elements are Floats}
\usage{
is.float(m)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}
}
\value{
a logical vector of the same length as \code{m}, indicating
whether each element is of the \code{float} sub-class.
}
\description{
Determine which Mooring Elements are Floats
}
\examples{
library(mooring)
m <- mooring(anchor(depth = 120), wire(length = 100), float("HMB 20"))
is.float(m)

}
\author{
Dan Kelley
}
