% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.instrument}
\alias{is.instrument}
\title{Determine which Mooring Elements are Instruments}
\usage{
is.instrument(m)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}
}
\value{
a logical vector of the same length as \code{m}, indicating
whether each element is of the \code{instruent} sub-class.
}
\description{
Determine which Mooring Elements are Instruments
}
\examples{
library(mooring)
m <- mooring(anchor(depth = 120), wire(length = 100), instrument("RD ADCP"),
    wire(length = 100), float("HMB 20"))
is.instrument(m)

}
\author{
Dan Kelley
}
