% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.misc}
\alias{is.misc}
\title{Determine which Mooring Elements are Misc}
\usage{
is.misc(m)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}
}
\value{
a logical vector of the same length as \code{m}, indicating
whether each element is of the \code{misc} sub-class.
}
\description{
Determine which Mooring Elements are Misc
}
\author{
Dan Kelley
}
