% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mooring}
\alias{is.mooring}
\title{Determine Whether an Object is a Mooring}
\usage{
is.mooring(m = NULL)
}
\arguments{
\item{m}{an object to be tested}
}
\description{
An object is a mooring if it is an unnamed list and if each
element in that list is a mooring element created by
e.g. \code{\link[=anchor]{anchor()}}, \code{\link[=wire]{wire()}}, or a related function.
For example, the
output of \code{\link[=mooring]{mooring()}} is a mooring, but the output of \code{\link[=anchor]{anchor()}} is not.
This function is mainly designed for use within the package that that, e.g.
\code{\link[=knockdown]{knockdown()}} will produce an error if its first argument is not a mooring.
}
