% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mooringElement}
\alias{is.mooringElement}
\title{Determine Whether an Object is a Mooring Element}
\usage{
is.mooringElement(m = NULL)
}
\arguments{
\item{m}{an object to be tested}
}
\description{
An object is a mooring element if it was created by
\code{\link[=anchor]{anchor()}}, \code{\link[=wire]{wire()}}, or a related function, each of
which returns an element of the \code{"mooringElement"} class.
}
