% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.release}
\alias{is.release}
\title{Determine which Mooring Elements are Release}
\usage{
is.release(m)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}
}
\value{
a logical vector of the same length as \code{m}, indicating
whether each element is of the \code{release} sub-class.
}
\description{
Determine which Mooring Elements are Release
}
\author{
Dan Kelley
}
