% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_wire.R
\name{isWire}
\alias{isWire}
\title{Indicate which mooring elements are wires}
\usage{
isWire(m)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}
}
\value{
a logical vector of the same length as \code{m}, indicating
whether each element is of the \code{wire} class.
}
\description{
Indicate which mooring elements are wires
}
\examples{
library(mooring)
m <- mooring(anchor(depth=120), wire(length=100), float("HMB 20"))
isWire(m)

}
\author{
Dan Kelley
}
