% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{knockdown}
\alias{knockdown}
\title{Compute how a mooring is knocked down by a current}
\usage{
knockdown(m, u = 1, debug = 0L)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class, usually created with
\code{\link[=discretise]{discretise()}}.}

\item{u}{either a non-negative number indicating depth-independent velocity,
or a function that returns that velocity as a function of depth (m)
below the ocean surface.}

\item{debug}{an integer value (or a numeric that can be converted to an
integer) that controls debugging. If this is positive, then some debugging
information may be printed by the function.}
}
\description{
The current may be a depth-independent or depth-dependent,
as specified by the \code{v} argument.
}
\examples{
# Illustrate importance of drag on the wire.
library(mooring)
m <- mooring(anchor(depth=100), wire(length=80), float("HMB 20"))
md <- discretise(m)
par(mfrow=c(1, 3))

# Example 1: no current
plot(md)

# Example 2: uniform 0.5 m/s (approx. 1 knot) current
k1 <- knockdown(md, u=0.5)
plot(k1, title="uniform 0.5 m/s")

# Example 3: 0.5 m/s current but only in top 40m of water column
k2 <- knockdown(md, u=function(depth) ifelse(depth < 40, 0.5, 0))
plot(k2, title="0.5 m/s only in top 40m")

}
\author{
Dan Kelley
}
