% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\title{Create a misc object}
\usage{
misc(
  model = "AanderaaT.chain",
  buoyancy = NULL,
  height = NULL,
  area = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the type of misc.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If this is \code{"?X"}, where X is a set of
characters, then \code{\link[=findElement]{findElement()}} is called to do a fuzzy search,
with the \code{search} argument set to \code{"misc"}.
(3) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(4) Otherwise, a new misc object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancy}{numeric value indicating buoyancy.  For
wire elements, this is buoyancy per length, expressed in kg/m.
For other elements, it is buoyancy, expressed in kg.
Note that \code{buoyancy} is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height (in m) of the element.
This value is used in the computation of the mooring geometry.
In some cases, as e.g. in Dewey-derived data that have "diameter" tabulated
as zero, it is also used in the computation of the area
of the element, as height*width.
Note that \code{height} is ignored if \code{model} is recognized.}

\item{area}{a numerical value indicating the horizontal instrument area
in \eqn{m^2}{m^2}.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}
}
\value{
\code{misc} returns an object of the \code{"mooring"} class and \code{"misc"} subclass.
}
\description{
Create a miscellaneous object,
either by looking up a known object from the database, or by defining a new type.
The function name, and most of the built-in data
values, come from Dewey's (1999, 2021) database.
For data derived from the Dewey database,
area is computed as in Dewey's Matlab code: if tabulated diameter is zero,
then the product of tabulated height and width is used; otherwise,
pi*(diameter/2)^2 is used.
}
\section{Data sources}{


To learn more about the data used for the built-in
misc
objects, use e.g.
\preformatted{
data(mooringElements)
mooringElements$miscs
}
}

\examples{
library(mooring)
# List known misc types
misc("?")

}
\references{
Dewey, Richard K. "Mooring Design & Dynamics-a Matlab® Package for
Designing and Analyzing Oceanographic Moorings." Marine Models, vol. 1, no. 1
(December 1, 1999): 103–57. https://doi.org/10.1016/S1369-9350(00)00002-X

Dewey, Richard. "Mooring Design and Dynamics:
A Matlab Package for Designing and Testing
Oceanographic Moorings And Towed Bodies."
Accessed May 15, 2021.
http://canuck.seos.uvic.ca/rkd/mooring/mdd/mdd.php
http://canuck.seos.uvic.ca/rkd/mooring/moordyn.php
}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{clamped}()},
\code{\link{connector}()},
\code{\link{float}()},
\code{\link{instrument}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
