% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\docType{package}
\name{mooring-package}
\alias{mooring-package}
\title{mooring: A Package for Analysing Oceanographic Moorings}
\description{
The mooring package provides functions for working with
oceanographic moorings.  The example provides a good
starting point for learning, and the documentation
for the key function, \code{\link[=mooring]{mooring()}}, provides the
next logical step in learning, after which the
vignettes should be consulted.
}
\examples{
library(mooring)
# Illustrate the deformation of a 100-m mooring in a 0.5 m/s
# (roughly 1 knot) current. Buoyancy is provided with a float
# of diameter 20 inches.
m <- mooring(anchor(depth=120), wire(length=100), float("HMB 20"))
par(mfrow=c(1, 3))
plot(m)
# Must discretise the wire portion to resolve the shape.
md <- discretise(m)
mdk <- knockdown(md, u=0.5)
plot(mdk)

}
