% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\docType{package}
\name{mooring}
\alias{mooring}
\title{mooring: A Package for Analysing Oceanographic Moorings.}
\description{
The mooring package provides functions for working with
oceanographic moorings.
}
\examples{
library(mooring)
m <- anchor(depth=100) + wire(length=80) + float("HMB 20")
print(m)
plot(m)

}
