% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{mooringDebug}
\alias{mooringDebug}
\title{Print a debugging message}
\usage{
mooringDebug(debug, v, ..., overview = FALSE, round = FALSE)
}
\arguments{
\item{debug}{an integer that controls debugging. If this is positive,
then the other arguments are given to \code{\link[=cat]{cat()}}, to print a message. Note
that no newline is printed, unless included in the \code{...} arguments.}

\item{v}{a value to be displayed.}

\item{...}{optional extra values to be displayed.}

\item{overview}{logical value indicating whether to summarize \code{v} (ignoring
\code{...}), by displaying its name and some of its values.}

\item{round}{either FALSE, for no rounding, or an integer, indicating the
rounding to be used, via a call to \code{\link[=round]{round()}}. This applies only to the case
where \code{overview} is TRUE.}
}
\description{
A function used for debugging within the package.
}
\author{
Dan Kelley
}
