% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{plot.mooring}
\alias{plot.mooring}
\title{Plot a mooring object}
\usage{
\method{plot}{mooring}(m)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class.}
}
\description{
Plot a mooring object
}
\author{
Dan Kelley
}
