% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\name{plot}
\alias{plot}
\title{Draw a mooring diagram}
\arguments{
\item{x}{a \code{mooring} object, created with \code{\link[=mooring]{mooring()}}.}

\item{which}{character value indicating the desired plot, with
choices: \code{"shape"} (the default), \code{"knockdown"}, \code{"tension"}
and \code{"velocity"}.}

\item{showInterfaces}{logical value indicating whether to indicate the water
surface with a blue line and the ocean bottom with a brown line.}

\item{showDepths}{logical value indicating whether to indicate the depths of
floats, to the left of the symbols.}

\item{showLabels}{logical value indicating whether to indicate anchors,
instruments and floats with \code{A}, \code{I} and \code{F}, respectively.}

\item{showDetails}{logical value indicating whether to show details for anchors,
instruments and floats with text.}

\item{fancy}{logical value indicating whether to indicate the
water and sediments with filled rectangles.  The alternative
is a simpler plot.}

\item{title}{character value indicating a title to put above
the plot.}

\item{mar}{numeric vector of length 4, used to set margins
outside axes. The default values reduce whitespace.
See \code{\link[=par]{par()}}.}

\item{mgp}{numeric vector of length 3, with a default that
reduces axis whitespace. See \code{\link[=par]{par()}}.}

\item{xaxs, yaxs}{character values that control axis ranges.
See \code{\link[=par]{par()}}.}

\item{xlim}{optional numeric vector of length 2 that can
be used to specify the limits of the horizontal axis.}

\item{type}{character value indicating type of plot. The default, \code{"l"},
means to draw lines, while e.g. \code{"p"} means to draw points.}

\item{debug}{an integer value (or a numeric that can be converted to an
integer) that controls debugging. If this is positive, then some debugging
information may be printed by the function.}

\item{...}{ignored}
}
\description{
Plot a side-view diagram that indicates mooring geometry.
}
\examples{
# Create, summarize, and plot a simple mooring comprising
# a bottom anchor, a 100-metre wire, and a float, in
# a current of 1 m/s.
library(mooring)
m <- mooring(anchor(), wire(length = 80), float("HMB 20"), waterDepth = 100) |>
    segmentize() |>
    knockdown(u = 1)
plot(m)
plot(m, which = "tension")

}
\author{
Dan Kelley
}
