% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{release}
\alias{release}
\title{Create a mooring-release object}
\usage{
release(model = "default_release")
}
\arguments{
\item{model}{character value indicating the model of the acoustic release.}
}
\value{
an object of the \code{"mooring"} class, with \code{type} equal to \code{"release"}.
}
\description{
Create a mooring-release object
}
\seealso{
Other functions that create mooring objects: 
\code{\link{+.mooring}()},
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{float}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
