% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{server}
\alias{server}
\title{Server for app, with standard arguments.}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{A list created by the shiny server, with entries for slider settings, etc.}

\item{output}{A list of output entries, for plotting, etc.}

\item{session}{A list used for various purposes.}
}
\description{
Server for app, with standard arguments.
}
\author{
Dan Kelley
}
