% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{[.mooring}
\alias{[.mooring}
\title{Access something in a mooring}
\usage{
\method{[}{mooring}(m, i)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{i}{either (a) an integer specifying index of item to be returned, or
(b) a character string. The first case is used to look up components of
a mooring. The second case requires \code{i} to be \code{"area"}, \code{"buoyancy"},
\code{"CD"}, or \code{"height"}, and the result is a single value if \code{m} is
an elementary object (example 1) or a whole mooring (example 2).}
}
\description{
Retrieves values from (a) a mooring element, as created with
\code{\link[=float]{float()}} or a similar function, or (b) a whole mooring, as created
with \code{\link[=mooring]{mooring()}}.
}
\examples{
library(mooring)
F <- float("HMB 20")
F["buoyancy"]
m <- mooring(anchor(depth = 120), wire(length = 100), F)
m["buoyancy"]

}
\author{
Dan Kelley
}
