% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mooring::mooring}
\alias{summary.mooring::mooring}
\alias{summary.mooring}
\title{Summarize a mooring}
\usage{
\method{summary}{`mooring::mooring`}(object, ...)
}
\arguments{
\item{object}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{...}{ignored.}
}
\description{
Summarize a mooring
}
\examples{
library(mooring)
m <- mooring(anchor(), wire(length = 80), float(), waterDepth = 100)
summary(m)

}
\author{
Dan Kelley
}
