% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{summary.mooring}
\alias{summary.mooring}
\title{Summarize a mooring}
\usage{
\method{summary}{mooring}(object, ...)
}
\arguments{
\item{object}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{...}{ignored.}
}
\description{
Summarize a mooring
}
\examples{
library(mooring)
# Simple case
m <- mooring(anchor(depth=100), wire(length=80), float("HMB 20"))
summary(m)
# Illustrate how it collects wire subintervals
md <- discretise(m)
mdk <- knockdown(md, 0.5)
summary(mdk)

}
\author{
Dan Kelley
}
