% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\name{summary}
\alias{summary}
\title{Summarize a mooring or a mooringElement}
\arguments{
\item{x}{a \code{mooring} object, created with \code{\link[=mooring]{mooring()}},
or a \code{moringElement} object, created with \code{\link[=anchor]{anchor()}}, \code{\link[=wire]{wire()}},
or any of the other functions that create mooring elements.}
}
\description{
Summarize a mooring or a mooringElement
}
\examples{
library(mooring)
a <- anchor()
a
m <- mooring(anchor(), wire(length = 80), float("HMB 20"), waterDepth = 100)
m

}
\author{
Dan Kelley
part of the item.
}
