% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summaryMooring}
\alias{summaryMooring}
\title{Summarize a mooring}
\usage{
summaryMooring(object, ...)
}
\arguments{
\item{object}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{...}{ignored.}
}
\description{
Summarize a mooring
}
\examples{
library(mooring)
m <- mooring(anchor(), wire(length = 80), float(), waterDepth = 100)
summaryMooring(m)

}
\author{
Dan Kelley
}
