% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tension.R
\name{tension}
\alias{tension}
\title{Tension between mooring elements}
\usage{
tension(m, stagnant = FALSE)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{stagnant}{logical value indicating whether to return
values for the mooring configuration in the stagnant case
(with zero velocity) or the present case (possibly with
nonzero velocity).}
}
\value{
a numeric vector of tension, in kg.
}
\description{
The first element (for the anchor) is repeated, so that the
length of the returned result matches the length of \code{m}.
}
\examples{
library(mooring)
m <- mooring(anchor(), wire(length = 100), float("HMB 20"), waterDepth = 120)
ms <- segmentize(m)
msk <- knockdown(ms, u = 0.5)
depth <- depth(msk)
# The right-hand panel indicates the anchor limit in red.
par(mfrow = c(1, 2))
plot(tension(msk), depth, ylim = rev(range(depth)), type = "l")
plot(msk, which = "tension")

}
\author{
Dan Kelley
}
