% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{wire}
\alias{wire}
\title{Create a wire object}
\usage{
wire(
  model = "1/4 wire/jack",
  buoyancy = NULL,
  length = NULL,
  width = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the model name of the wire. The
default, \code{"1/4 wire/jack"}, is for quarter-inch jacketed steel wire. If
model is set to \code{"?"}, then the default is returned, but first a list
of possible model values is printed. If model is not recognized, then
\link{wire} creates a new type, using the stated diameter and CD values;
otherwise, these arguments are ignored.}

\item{buoyancy}{numeric value of the buoyancy per length of wire, in kg/m.}

\item{length}{numeric value indicating the length of the wire, in metres.}

\item{width}{numeric value for the width of the wire, in m.
This is ignored if \code{model} is recognized.}

\item{CD}{numeric value for the drag coefficient for the wire.
This is ignored if \code{model} is recognized.}
}
\value{
\link{wire} returns an object of the \code{"mooring"} class, with \code{type} equal to \code{"wire"}.
}
\description{
Creates an object that describes a mooring wire, or other type of line, e.g.
nylon ropes are considered wires, for this purpose.
Note that all dimensions are specified in m, not cm.
}
\details{
The permitted \code{model} values are listed alphabetically below, as listed with \code{wire("?")}:
\preformatted{
"1 Nylon"          "1/2 AmSteel-Blue" "1/2 Dacron"
"1/2 Dyneema"      "1/2 Kevlar"       "1/2 Neutral Wire"
"1/2 VLS"          "1/2 wire/jack"    "1/4 AmSteel-Blue"
"1/4 Kevlar"       "1/4 wire rope"    "1/4 wire/jack"
"1/4-1/2 WarpSpd"  "11mm Perlon"      "1in Neutral Wire"
"3/16 wire rope"   "3/4 Dacron"       "3/4 Nylon"
"3/4 Polyprop"     "3/8 AmSteel-Blue" "3/8 Kevlar"
"3/8 wire rope"    "3/8 wire/jack"    "5/16 Kevlar"
"5/16 wire rope"   "5/16 wire/jack"   "5/8 AmSteel-Blue"
"5/8 Kevlar"       "5/8 Nylon"        "7/16 Dacron"
"7/16 Kevlar"      "7/16 VLS"         "9/16 Dacron"
"9/16 Kevlar"      "BPS Cable"        "BPS Power/Coms"
}
}
\seealso{
Other functions that create mooring objects: 
\code{\link{+.mooring}()},
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{float}()},
\code{\link{release}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
