% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{x}
\alias{x}
\title{Return horizontal coordinate, in m, of elements in mooring.}
\usage{
x(m)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class.}
}
\value{
a numeric vector of horizontal coordinate in metres.
}
\description{
Return horizontal coordinate, in m, of elements in mooring.
}
\examples{
library(mooring)
m <- anchor(depth=120) + wire(length=100) + float("HMB 20")
x(m)

}
\author{
Dan Kelley
}
