% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{z}
\alias{z}
\title{Return vertical coordinate, in m, of elements in mooring.}
\usage{
z(m)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class.}
}
\value{
a numeric vector of vertical coordinate in metres.
}
\description{
This is the negative of \code{\link[=depth]{depth()}}.
}
\examples{
library(mooring)
m <- anchor(depth=120) + wire(length=100) + float("HMB 20")
z(m)

}
\author{
Dan Kelley
}
