% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{anchor}
\alias{anchor}
\title{Create an anchor object}
\usage{
anchor(model = "default_anchor")
}
\arguments{
\item{model}{character value indicating the model of the anchor.}
}
\value{
an object of the \code{"mooring"} class, with \code{type} equal to \code{"anchor"}.
}
\description{
Create an anchor object
}
\seealso{
Other functions that create mooring objects: 
\code{\link{cable}()},
\code{\link{float}()},
\code{\link{release}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
