% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor.R
\name{anchorWeight}
\alias{anchorWeight}
\title{Find Anchor Wweight of a Mooring}
\usage{
anchorWeight(m)
}
\arguments{
\item{m}{an object of class \code{"mooring"}, created with \code{\link[=mooring]{mooring()}}.}
}
\value{
\code{anchorWeight} returns the weight (i.e. the negative of the
buoyancy) of the anchor in the mooring.
}
\description{
Find Anchor Wweight of a Mooring
}
\author{
Dan Kelley
}
