% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{app}
\alias{app}
\title{Run a GUI app for interactive simulations}
\usage{
app(options = list(height = 500))
}
\arguments{
\item{options}{List containing options that are provided
to \code{\link[shiny]{shinyApp}}, which creates the GUI app.}
}
\description{
This makes a simple mooring with an anchor, a line, and a float, in
a depth-uniform current.  Sliders adjust line length and current speed.
Pulldown menus adjust wire and float types.
}
\author{
Dan Kelley
}
