% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app2bs.R
\name{app2bs}
\alias{app2bs}
\title{Run a two-element GUI app for interactive simulations}
\usage{
app2bs(debug = FALSE)
}
\arguments{
\item{debug}{logical value indicating whether to print debugging
output.}
}
\description{
This makes a simple mooring with an anchor, an instrument, and
a float, in a variety of current profiles. It differs from \code{\link[=app1]{app1()}}
in having the instrument, and in having some interface changes.
It differs from \code{\link[=app2]{app2()}} in using the 'bslib' GUI framework,
which is cleaner than the framework used in \code{\link[=app1]{app1()}} and \code{\link[=app2]{app2()}}.
}
\seealso{
Other interactive apps: 
\code{\link{app1}()},
\code{\link{app2}()}
}
\author{
Dan Kelley
}
\concept{interactive apps}
