% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Get Mooring/Element side-view Area}
\usage{
area(m, phi = TRUE)
}
\arguments{
\item{m}{either a mooring object, created by \code{\link[=mooring]{mooring()}}, or a
mooring-element object, created with \code{\link[=anchor]{anchor()}}, \code{\link[=chain]{chain()}},
\code{\link[=connector]{connector()}}, \code{\link[=instrument]{instrument()}}, \code{\link[=misc]{misc()}}, \code{\link[=release]{release()}}, or \code{\link[=wire]{wire()}}.}

\item{phi}{a numeric or logical value that controls how the angle
is computed.  If this is FALSE, no angle is used.  If it is TRUE,
which is the default, then the value of \code{phi} within the mooring
element is used (or 0 is used, if there is no value).  Finally, if
\code{phi} is a single numerical value, then that is taken as the angle
to the vertical, in radians.}
}
\value{
\code{area} returns a numeric value of the side-view area, in m^2.
}
\description{
\code{area} returns a numeric vector of mooring element side-view
area(s), in square metres, taking into account the orientation of
the element with respect to the vertical. (Floats are assumed to be
spherical, so the orientation is irrelevant. This is not the case
for wires and instruments.) The return value will be a single
number if \code{m} is a single element but in a typical mooring it will
have multiple values. The areas are listed with the top element
first, i.e. in the reverse order to that used in constructing the
mooring with \code{\link[=mooring]{mooring()}}.
}
\details{
When initially created, objects do not contain an angle value, but
such values are inserted by \code{\link[=knockdown]{knockdown()}}. Moorings that are
significantly tilted will have reduced side-view area, which
will reduce the current-induced drag, as computed with \code{\link[=drag]{drag()}}.
}
\examples{
library(mooring)
# Floats are unaffected by rotation
area(float())
area(float(), pi / 4)
# Other elements, e.g. instruments, are affected by rotation
area(instrument())
area(instrument(), pi / 4.0) * sqrt(2.0)


}
\author{
Dan Kelley
}
