% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{area}
\alias{area}
\title{Area of mooring elements}
\usage{
area(m)
}
\arguments{
\item{m}{either a mooring object, created by \code{\link[=mooring]{mooring()}}, or a
mooring-element object, created with \code{\link[=anchor]{anchor()}}, \code{\link[=chain]{chain()}}, \code{\link[=release]{release()}},
\code{\link[=wire]{wire()}}, \code{\link[=instrument]{instrument()}}, or \code{\link[=float]{float()}}.}
}
\value{
\code{area} returns a numeric value of the area viewed from a horizontal, in m^2.
}
\description{
Area of mooring elements
}
\examples{
library(mooring)
area(float())

}
\author{
Dan Kelley
}
