% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{area}
\alias{area}
\title{Get element areas}
\usage{
area(m)
}
\arguments{
\item{m}{either a multi-element mooring, created by \code{\link[=mooring]{mooring()}}, or a mooring element,
created with e.g. \code{\link[=anchor]{anchor()}}, \code{\link[=chain]{chain()}}, \code{\link[=release]{release()}}, \code{\link[=wire]{wire()}}, \code{\link[=instrument]{instrument()}},
or \code{\link[=float]{float()}}.}
}
\description{
Get element areas
}
\author{
Dan Kelley
}
