% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoyancy.R
\name{buoyancy}
\alias{buoyancy}
\title{Buoyancy of mooring elements}
\usage{
buoyancy(m, debug = 0L)
}
\arguments{
\item{m}{either a mooring object, created by \code{\link[=mooring]{mooring()}}, or a
mooring-element object, created with}

\item{debug}{an integer value (or a numeric that can be converted to an
integer) that controls debugging. If this is positive, then some debugging
information may be printed by the function.}
}
\value{
\code{buoyancy} returns a numeric vector of buoyancy, expressed
in kg or N, depending on the value of \code{unit}.
}
\description{
This returns 'buoyancy' in kg, as is the convention of storage in
this package (and is quite common in practical work).  It must be
multiplied by the acceleration due to gravity, g=9.81m/s^2, to get
an \emph{actual} buoyancy force, in Newtons. This multiplication is done
in \code{\link[=knockdown]{knockdown()}} and \code{\link[=tension]{tension()}}.
}
\details{
Note that the present version of this function does not account for
depth variations in seawater density, for those tend to be well
under 1 percent, and other forces involved in mooring dynamics are
much more uncertain than that.  For example, Hamilton (1989) found
that oscillations in mooring lines could lead to enhanced drag, in
some cases necessitating an increase in CD for wire from 1.4 to 2.6
(see captions of his figures 12 and 13).
}
\examples{
library(mooring)
m <- mooring(anchor(depth = 120), wire(length = 100), float("HMB 20"))
buoyancy(m)

}
\references{
Hamilton, J. M. "The Validation and Practical Applications of a Sub-Surface
Mooring Model." Canadian Technical Report of Hydrography and Ocean
Sciences. Bedford Institute of Oceanography, 1989.
}
\author{
Dan Kelley
}
