% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoyancy.R
\name{buoyancy}
\alias{buoyancy}
\title{Buoyancy of mooring elements}
\usage{
buoyancy(m, debug = 0L)
}
\arguments{
\item{m}{either a mooring object, created by \code{\link[=mooring]{mooring()}}, or a
mooring-element object, created with \code{\link[=anchor]{anchor()}}, \code{\link[=chain]{chain()}},
\code{\link[=connector]{connector()}}, \code{\link[=instrument]{instrument()}}, \code{\link[=misc]{misc()}}, \code{\link[=release]{release()}}, or \code{\link[=wire]{wire()}}.}

\item{debug}{an integer value (or a numeric that can be converted to an
integer) that controls debugging. If this is positive, then some debugging
information may be printed by the function.}
}
\value{
\code{buoyancy} returns a numeric vector of buoyancy, expressed
in kg or N, depending on the value of \code{unit}.
}
\description{
This returns 'buoyancy' in kg, as is the convention of storage in
this package (and is quite common in practical work).  It must be
multiplied by the acceleration due to gravity, g=9.81m/s^2, to get
an \emph{actual} buoyancy force, in Newtons. This multiplication is done
in \code{\link[=knockdown]{knockdown()}} and \code{\link[=tension]{tension()}}.
}
\details{
Note that the present version of this function does not account for
depth variations in seawater density, for those tend to be well
under 1 percent.
}
\examples{
library(mooring)
m <- mooring(anchor(), wire(length = 100), float("HMB 20"), waterDepth = 120)
buoyancy(m)

}
\author{
Dan Kelley
}
