% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{CD}
\alias{CD}
\title{CD of mooring elements}
\usage{
CD(m)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}
}
\description{
This looks up element \code{CD} values for the elements
in a mooring.  For anchor items, which are assumed to
have no drag, it fills in 0.
}
\author{
Dan Kelley
}
