% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD.R
\name{CD}
\alias{CD}
\title{Get mooring/element drag coefficient}
\usage{
CD(m)
}
\arguments{
\item{m}{either a mooring object, created by \code{\link[=mooring]{mooring()}}, or a
mooring-element object, created with \code{\link[=anchor]{anchor()}}, \code{\link[=chain]{chain()}},
\code{\link[=connector]{connector()}}, \code{\link[=instrument]{instrument()}}, \code{\link[=misc]{misc()}}, \code{\link[=release]{release()}}, or \code{\link[=wire]{wire()}}.}
}
\value{
\code{CD} returns a numeric vector of drag coefficient(s).
}
\description{
Get mooring/element drag coefficient
}
\examples{
library(mooring)
CD(float())

}
\author{
Dan Kelley
}
