% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{chain}
\alias{chain}
\title{Create a chain object}
\usage{
chain(
  model = "1in buoy chain",
  buoyancyPerMeter = NULL,
  width = NULL,
  CD = NULL,
  length = NULL
)
}
\arguments{
\item{model}{character value indicating the type of chain.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(3) Otherwise, a new chain object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancyPerMeter}{numeric value indicating buoyancy per meter
of length, expressed as kg/m.
Note that \code{buoyancyPerMeter} is ignored if \code{model} is recognized.}

\item{width}{numeric value for the width (in m)
of the element as viewed from above.
This is multiplied by \code{height} to compute the
area of the element, as part of the drag computation.
Note that \code{width} is ignored if \code{model} is recognized.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}

\item{length}{(mandatory) numeric value indicating the length (in m) of the wire.}
}
\value{
\code{chain} returns an object of the \code{"mooring"} class and \code{"chain"} subclass.
}
\description{
Create an object that describes mooring chain elements such as shackles,
either by looking up a known object from the database, or defining a new type.
}
\examples{
library(mooring)
# List known chain types
chain("?")

}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{connector}()},
\code{\link{float}()},
\code{\link{instrument}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
