% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{chain}
\alias{chain}
\title{Create a chain object}
\usage{
chain(
  model = "1\\" buoy chain",
  buoyancy = NULL,
  height = NULL,
  width = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the model name of the chain. The
default is \verb{"1\\" buoy chain"}.
Use \code{chain("?")} to get a listing of permitted values.
If \code{model} is not recognized, then
\link{chain} creates a new type, using the stated diameter and CD values;
otherwise, these arguments are ignored.}

\item{buoyancy}{numeric value of the buoyancy of the item, in kg.
This is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height of the element, in m.
This is ignored if \code{model} is recognized.}

\item{width}{numeric value for the width of the element, in m.
This is ignored if \code{model} is recognized.}

\item{CD}{numeric value for the drag coefficient for the item.
This is ignored if \code{model} is recognized.}
}
\value{
\link{chain} returns an object of the \code{"mooring"} class, with \code{type} equal to \code{"chain"}.
}
\description{
Creates an object that describes mooring chain elements such as shackles.
Note that all dimensions are specified in m, not cm.
}
\examples{
library(mooring)
# List known chain types
chain("?")

}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{float}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
