% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamped.R
\name{clamped}
\alias{clamped}
\title{Clamp an Instrument on the Wire}
\usage{
clamped(instrument)
}
\arguments{
\item{instrument}{a \code{mooring} object of subtype \code{instrument},
typically as created with \code{\link[=instrument]{instrument()}}.}
}
\value{
\code{clamped} returns an object of the \code{"mooring"} class and \code{"instrument"} subclass.
}
\description{
This changes the \code{height} field of an \link{instrument} object to 0,
so that it does not contribute to the overall length of the mooring.
Using this in deep water is unlikely to make much difference,
but it is handy for shallow-water cases, if the device was
clamped on to the wire, but the \code{height} field of the
\code{instrument} object used to represent it has non-zero \code{height}.
The example illustrates this, for mooring in very shallow water.
}
\examples{
library(mooring)
a <- function(depth) {
    anchor("my anchor",
        buoyancy = -50, height = 0.1, CD = 0,
        depth = depth
    )
}
w <- function(length) wire(length = length)
i <- instrument("SBE37 microcat clamp-on style")
f <- float("my float", buoyancy = 20, height = 0.2, area = 0.2^2, CD = 1.3)
# Construct unclamped mooring (m) and clamped mooring (M).
m <- mooring(a(2), w(1), i, w(0.2), f) |>
    discretise(by = 0.1) |>
    knockdown(u = 1)
M <- mooring(a(2), w(1), clamped(i), w(0.2), f) |>
    discretise(by = 0.1) |>
    knockdown(u = 1)
# Plot unclamped case (top row) and clamped case (bottom row).
par(mfrow = c(2, 2))
plot(m)
plot(m, which = "tension")
# Bottom row: clamped version
plot(M)
plot(M, which = "tension")

}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{connector}()},
\code{\link{float}()},
\code{\link{instrument}()},
\code{\link{misc}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
