% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{depth}
\alias{depth}
\title{Return depth, in m, of elements in mooring.}
\usage{
depth(m)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class.}
}
\value{
a numeric vector of depth in metres.
}
\description{
This is the negative of \code{\link[=z]{z()}}.
}
\examples{
library(mooring)
m <- anchor(depth=120) + wire(length=100) + float("HMB 20")
depth(m)

}
\author{
Dan Kelley
}
