% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{discretise}
\alias{discretise}
\title{Discretise the chain and wire portions of a mooring}
\usage{
discretise(m, by = 1)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{by}{numeric value giving the rough size of the chunks.
The actual size is computed as the length of wire, divided
by the rounded ratio of that length to \code{by}. For example,
using \code{by=10} with a 95-m length of wire will result
in chunks of length 9.5m, not 10m.  In shallow water moorings,
the default value of 1m makes sense, but larger values
might be employed for moorings in the deep ocean. If \code{by}
exceeds the height of a wire portion, then that portion is not
subdivided.}
}
\value{
an object of the \code{"mooring"} class, identical
to \code{m} except that wire portions are chopped up into shorter
pieces.
}
\description{
Break up \code{chain} and \code{wire} portions of a mooring into smaller chunks,
so that the deformation by a current can be traced more
accurately by \code{\link[=knockdown]{knockdown()}}.
}
\author{
Dan Kelley
}
