% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{drag}
\alias{drag}
\title{Drag on mooring elements}
\usage{
drag(m, u, rho = 1027, g = 9.8)
}
\arguments{
\item{m}{either a mooring object, created by \code{\link[=mooring]{mooring()}}, or a
mooring-element object, created with}

\item{u}{either a non-negative number indicating depth-independent velocity,
or a function that returns that velocity as a function of depth (m)
below the ocean surface.}

\item{rho}{a numeric value giving
water density in kg/m^3. The default value of 1027 is
likely to be sufficient for most calculations, given that
oceanographic variations of density are typically well under 1
percent, although it might make sense to use a smaller value
in lake applications.  Note that the overall force balance
is more sensitive to other things, as discussed in the \dQuote{Details}
section of the documentation for \code{\link[=buoyancy]{buoyancy()}}.}

\item{g}{numeric value of the acceleration due to gravity, with default
being 9.8 m/s^2.}
}
\value{
\code{drag} returns a numeric vector of horizontal drag "force" (really, force
divided by gravitational acceleration), expressed in kg.
}
\description{
This looks up element areas with \code{\link[=area]{area()}} and drag
coefficients with \code{\link[=CD]{CD()}}, then computes drag
force (in Newtons) with
\eqn{(1/2)*area*rho*CD*u^2}{(1/2)*area*rho*CD*u^2}
and divides by \code{g} to get a mass equivalance, which
is returned.
}
\details{
Although fluid density \code{rho} and \code{g} are parameters to this
function, the default values are likely to be used in all
practical oceanographic calculations, because drag coefficient
is not known to three digits.
}
\author{
Dan Kelley
}
