% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{findElement}
\alias{findElement}
\title{Find a mooring element with a fuzzy search}
\usage{
findElement(e, max.distance = 0.1)
}
\arguments{
\item{e}{character value to be used for the fuzzy match, passed on to \code{\link[=agrep]{agrep()}}.}

\item{max.distance}{numeric value passed on to \code{\link[=agrep]{agrep()}}.}
}
\value{
\code{findElement} returns (silently) a list of the suggested function calls.
}
\description{
\code{findElement} does a fuzzy search for an element model, using
\code{\link[=agrep]{agrep()}}.  The output (if any) is in the form of suggested calls
to element-creating functions
\code{\link[=anchor]{anchor()}}, \code{\link[=chain]{chain()}}, \code{\link[=connector]{connector()}}, \code{\link[=float]{float()}}, \code{\link[=instrument]{instrument()}}, and \code{\link[=wire]{wire()}}.
}
\examples{
library(mooring)
findElement("nylon")

}
\author{
Dan Kelley
}
