% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{float}
\alias{float}
\title{Create a float object}
\usage{
float(model = "Hydro Float 20")
}
\arguments{
\item{model}{character value indicating the model of the float.}
}
\value{
an object of the \code{"mooring"} class, with \code{type} equal to \code{"float"}.
}
\description{
Create a float object
}
\seealso{
Other functions that create mooring objects: 
\code{\link{+.mooring}()},
\code{\link{anchor}()},
\code{\link{cable}()},
\code{\link{release}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
