% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{float}
\alias{float}
\title{Create a float object}
\usage{
float(
  model = "Kiel SFS40in",
  buoyancy = NULL,
  height = NULL,
  diameter = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the type of float.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(3) Otherwise, a new float object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancy}{numeric value indicating buoyancy.  For
wire elements, this is buoyancy per length, expressed in kg/m.
For other elements, it is buoyancy, expressed in kg.
Note that \code{buoyancy} is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height (in m) of the element.
This value is used in the computation of (a) the mooring geometry
and (b) the area of the element (which is in turn used in
the drag computation).
Note that \code{height} is ignored if \code{model} is recognized.}

\item{diameter}{numeric value for the diameter (in m)
of the element as viewed from a horizontal direction.
This is used to compute area, for drag calculations, as
pi*(diameter/2)^2.
Note that \code{diameter} is ignored if \code{model} is recognized.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}
}
\value{
\code{float} returns an object of the \code{"mooring"} class and \code{"float"} subclass.
}
\description{
Create a float object,
either by looking up a known object from the database, or defining a new type.
}
\details{
Note that \code{HMB} in a name is a short-hand for \verb{Hydrofloat Mooring Buoy}.  Data for these
floats was extracted from Reference 1 on 2021-05-19 by Dan Kelley, with the \code{CD} value
being set at 0.65 (in the absence of any data from the manufacture) because that
value is used in Dewey's dataset for many floats.
}
\examples{
library(mooring)
# List known float types
float("?")

}
\references{
\enumerate{
\item \url{https://deepwaterbuoyancy.com/wp-content/uploads/hydro-float-mooring-buoys-deepwater-buoyancy.pdf}
}
}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{instrument}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
