% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/float.R
\name{float}
\alias{float}
\title{Create a float object}
\usage{
float(
  model = "Kiel SFS40in",
  buoyancy = NULL,
  height = NULL,
  area = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the type of float
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If this is \code{"?X"}, where X is a set of
characters, then \code{\link[=findElement]{findElement()}} is called to do a fuzzy search,
with the \code{search} argument set to \verb{"float (3) If it is recognized, i.e. if that type is stored in }data(mooringElements)`, then the that stored value
is used, and all other arguments to this function are ignored.
(4) Otherwise, a new float
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancy}{numeric value indicating buoyancy.  For
wire elements, this is buoyancy per length, expressed in kg/m.
For other elements, it is buoyancy, expressed in kg.
Note that \code{buoyancy} is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height (in m) of the element.
This value is used in the computation of the mooring geometry.
In some cases, as e.g. in Dewey-derived data that have "diameter" tabulated
as zero, it is also used in the computation of the area
of the element, as height*width.
Note that \code{height} is ignored if \code{model} is recognized.}

\item{area}{a numerical value indicating the horizontal instrument area
in \eqn{m^2}{m^2}.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}
}
\value{
\code{float} returns an object of the \code{"mooringElement"} class and \code{"float"} subclass.
}
\description{
Create a float object
}
\section{Data sources}{


To learn more about the data used for the built-in
float
objects, use e.g.
\preformatted{
data(mooringElements)
mooringElements$float
}
}

\examples{
library(mooring)
# List known float types
float("?")

}
\references{
\enumerate{
\item \url{https://deepwaterbuoyancy.com/wp-content/uploads/hydro-float-mooring-buoys-deepwater-buoyancy.pdf}
}
}
\seealso{
Other functions that create mooring elements: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{connector}()},
\code{\link{instrument}()},
\code{\link{misc}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring elements}
