% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{float}
\alias{float}
\title{Create a float object}
\usage{
float(
  model = "Kiel SFS40in",
  buoyancy = NULL,
  height = NULL,
  diameter = NULL,
  CD = NULL,
  source = NULL
)
}
\arguments{
\item{model}{character value indicating the model name of the float. The
default is \code{"Kiel SFS40in"}.
Use \code{float("?")} to get a listing of permitted values.
If \code{model} is not recognized, then
\link{float} creates a new type, using the stated diameter and CD values;
otherwise, these arguments are ignored.}

\item{buoyancy}{numeric value of the buoyancy of the float, in kg.
This is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height of the element, in m.
This is ignored if \code{model} is recognized.}

\item{diameter}{numeric value for the diameter of the float, in m.
This is ignored if \code{model} is recognized.}

\item{CD}{numeric value for the drag coefficient for the item.}

\item{source}{character value indicating the source of the data.}
}
\value{
\link{float} returns an object of the \code{"mooring"} class, with \code{type} equal to \code{"float"}.
}
\description{
Create a float object.  Some notes follow.
\enumerate{
\item \code{HMB} in a name is a short-hand for \verb{Hydrofloat Mooring Buoy}.  Data for these
floats was extracted from Reference 1 on 2021-05-19 by Dan Kelley, with the \code{CD} value
being set at 0.65 (in the absence of any data from the manufacture) because that
value is used in Dewey's dataset for many floats.
}
}
\examples{
library(mooring)
# List known float types
float("?")

}
\references{
\enumerate{
\item \url{https://deepwaterbuoyancy.com/wp-content/uploads/hydro-float-mooring-buoys-deepwater-buoyancy.pdf}
}
}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
