% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{height}
\alias{height}
\title{Get mooring/element height}
\usage{
height(m)
}
\arguments{
\item{m}{either a mooring object, created by \code{\link[=mooring]{mooring()}}, or a
mooring-element object, created with}
}
\value{
\code{height} returns a numeric vector of element(s) height, in m.  This will be
a single value if the first element is a single element, or a longer vector if it
is a mooring.
}
\description{
Get mooring/element height
}
\examples{
library(mooring)
height(float())
m <- mooring(anchor(depth = 120), wire(length = 100), float("HMB 20"))
height(m)

}
\author{
Dan Kelley
}
