% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\name{instrument}
\alias{instrument}
\title{Create an instrument object}
\usage{
instrument(
  model = "SBE37 microcat clamp-on style",
  buoyancy = NULL,
  height = NULL,
  area = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the type of instrument.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If this is \code{"?X"}, where X is a set of
characters, then \code{\link[=findElement]{findElement()}} is called to do a fuzzy search,
with the \code{search} argument set to \code{"instrument"}.
(3) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(4) Otherwise, a new instrument object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancy}{numeric value indicating buoyancy.  For
wire elements, this is buoyancy per length, expressed in kg/m.
For other elements, it is buoyancy, expressed in kg.
Note that \code{buoyancy} is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height (in m) of the element.
This value is used in the computation of the mooring geometry.
In some cases, as e.g. in Dewey-derived data that have "diameter" tabulated
as zero, it is also used in the computation of the area
of the element, as height*width.
Note that \code{height} is ignored if \code{model} is recognized.}

\item{area}{a numerical value indicating the horizontal instrument area
in \eqn{m^2}{m^2}.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}

\item{source}{character value indicating the source of the data.}
}
\value{
\code{instrument} returns a \code{"mooringElement"} object with \code{"instrument"} subclass.
}
\description{
Create an instrument object,
either by looking up a known object from the database, or by defining a new type.
}
\examples{
library(mooring)
# List known instrument types
instrument("?")

}
\seealso{
Other functions that create mooring elements: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{connector}()},
\code{\link{float}()},
\code{\link{misc}()},
\code{\link{release}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring elements}
