% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{knockdown}
\alias{knockdown}
\title{Compute how a mooring is knocked down by a current}
\usage{
knockdown(m, u = 1, debug = 0L)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class, usually created with
\code{\link[=discretise]{discretise()}}.}

\item{u}{either a non-negative number indicating depth-independent velocity,
or a function that returns that velocity as a function of the vertical
coordinate.  For the second case, note that the vertical coordinate is
defined as height (in m) above the sea surface, so that e.g.
\code{v=function(z) 0.5*exp(z/100)} mould be used to specify a current that
is 0.5m/s at the surface, dropping exponentially through the water column,
reaching to 0.18m/s a hundred metres below the surface.}

\item{debug}{an integer controlling debugging.  The default value of 0
means to work silently. Use a positive value to cause the function to
print some information about intermediate results.}
}
\description{
The current may be a depth-independent or depth-dependent,
as specified by the \code{v} argument.
}
\examples{
library(mooring)
# Illustrate importance of drag on the wire.
m <- anchor(depth=100) + wire(length=80) + float("HMB 20")
md <- discretise(m)
# No knockdown
plot(md)
# Knockdown in uniform 0.5 m/s current
k1 <- knockdown(md, u=1)
plot(k1)
# Knockdown in 0.5 m/s current but only in top 40m of water column
k2 <- knockdown(md, u=function(z) ifelse(z>-40, 1, 0))
plot(k2)

}
\author{
Dan Kelley
}
