% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{knockdown}
\alias{knockdown}
\title{Compute how a mooring is knocked down by a current}
\usage{
knockdown(m, u = 1, debug = 0L)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class, usually created with
\code{\link[=discretise]{discretise()}}.}

\item{u}{either a non-negative number indicating depth-independent velocity,
or a function that returns that velocity as a function of depth (m)
below the ocean surface.}

\item{debug}{an integer value (or a numeric that can be converted to an
integer) that controls debugging. If this is positive, then some debugging
information may be printed by the function.}
}
\value{
a new \code{mooring} object representing the deformed mooring, with
\code{x} and \code{z} values updated (and original values saved as \code{x0} and \code{z0}).
}
\description{
The current may be a depth-independent or depth-dependent,
as specified by the \code{u} argument.  The returned result has
an attribute named \code{u} that holds the value of that
argument, and this is how a later call to \code{\link[=plot.mooring]{plot.mooring()}}
is able to display a velocity profile; see
\dQuote{Examples} 2 and 3.
}
\examples{
# Illustrate importance of drag on the wire.
library(mooring)
m <- mooring(anchor(depth=100), wire(length=80), float("HMB 20"))
md <- discretise(m)

# Example 1: no current
plot(md)

# Example 2: uniform 0.5 m/s (approx. 1 knot) current
par(mfrow=c(1, 2))
k1 <- knockdown(md, u=0.5)
plot(k1, which="velocity")
plot(k1)

# Example 3: 0.5 m/s current at surface, decaying below
k2 <- knockdown(md, u=function(depth) 0.5*exp(-depth/100))
par(mfrow=c(1, 2))
plot(k2, which="velocity")
plot(k2)

}
\author{
Dan Kelley
}
