% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{knockdown}
\alias{knockdown}
\title{Compute how a mooring is knocked down by a current}
\usage{
knockdown(m, u = 1, debug = 0L)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class, usually created with
\code{\link[=discretise]{discretise()}}.}

\item{u}{either a non-negative number indicating depth-independent velocity,
or a function that returns that velocity as a function of depth (m)
below the ocean surface; see \dQuote{Examples}.}

\item{debug}{an integer controlling debugging.  The default value of 0
means to work silently. Use a positive value to cause the function to
print some information about intermediate results.}
}
\description{
The current may be a depth-independent or depth-dependent,
as specified by the \code{v} argument.
}
\examples{
library(mooring)
# Illustrate importance of drag on the wire.
par(mfrow=c(1, 3))
m <- anchor(depth=100) + wire(length=80) + float("HMB 20")
md <- discretise(m)
# No knockdown
plot(md)
# Knockdown in uniform 0.5 m/s (approx. 1 knot) current
k1 <- knockdown(md, u=0.5)
plot(k1, title="uniform 0.5 m/s")
# Knockdown in 0.5 m/s current but only in top 40m of water column
k2 <- knockdown(md, u=function(depth) ifelse(depth < 40, 0.5, 0))
plot(k2, title="0.5 m/s only in top 40m")

}
\author{
Dan Kelley
}
