% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{knockdown}
\alias{knockdown}
\title{Compute how a mooring is knocked down by a current}
\usage{
knockdown(m)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class, usually created with
\code{\link[=discretise]{discretise()}}.}
}
\description{
Compute how a mooring is knocked down by a current
}
\author{
Dan Kelley
}
