% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring-package.R
\docType{package}
\name{mooring-package}
\alias{mooring-package}
\title{mooring: Design and Model Oceanographic Moorings}
\description{
Calculates the response of oceanographic moorings to horizontal currents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dankelley.github.io/mooring/}
  \item Report bugs at \url{https://github.com/dankelley/mooring/issues}
}

}
\author{
\strong{Maintainer}: Dan Kelley \email{dan.kelley@dal.ca} (\href{https://orcid.org/0000-0001-7808-5911}{ORCID})

}
\keyword{internal}
