% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{mooring}
\alias{mooring}
\title{Create a mooring}
\usage{
mooring(..., waterDepth = NA)
}
\arguments{
\item{...}{two or more elementary objects, as created by
\code{\link[=anchor]{anchor()}}, \code{\link[=release]{release()}}, \code{\link[=chain]{chain()}},
\code{\link[=wire]{wire()}}, \code{\link[=connector]{connector()}}, \code{\link[=instrument]{instrument()}}, \code{\link[=misc]{misc()}} or \code{\link[=float]{float()}}.}

\item{waterDepth}{numerical value giving the full depth of
the water column.  For a sub-surface mooring, this must exceed
the sum of the heights of all the mooring elements.}
}
\value{
\code{mooring} returns an object in the \code{mooring} class that
contains an item named \code{elements}, which holds the items provided
in the \code{...} argument.  Each of those items is of the
\code{mooringElement} class.
}
\description{
Assemble elementary components into a mooring string.
The first argument must be an anchor, created
with \code{\link[=anchor]{anchor()}}.  These are stored in reverse order
in the return value (see Example 1).
}
\examples{
# Example 1: a simple mooring with a single instrument
library(mooring)
m <- mooring(anchor(),
    wire(length = 70), instrument(), wire(length = 10),
    float(), waterDepth = 100)
ms <- segmentize(m)
msk <- knockdown(ms, u = 1)
draw(msk)


}
\author{
Dan Kelley
}
