% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{mooring}
\alias{mooring}
\title{Create a mooring}
\usage{
mooring(...)
}
\arguments{
\item{...}{two or more elementary objects, e.g. as created by \code{\link[=anchor]{anchor()}},
\code{\link[=chain]{chain()}}, \code{\link[=wire]{wire()}}, or \code{\link[=float]{float()}}.}
}
\description{
Assemble components into a mooring, starting with an anchor, created with \code{\link[=anchor]{anchor()}},
and then adding wires, created with \code{\link[=wire]{wire()}}, instruments, created with \code{\link[=instrument]{instrument()}},
floats, created with \code{\link[=float]{float()}}, etc.
}
\examples{
# Example 1: most basic case: anchor, line, float.
library(mooring)
m <- mooring(anchor(depth=100), wire(length=80), float("HMB 20"))

# Example 2: mooring with mid-depth instrument and a float
# just below it.  Note the definition of some functions,
# to simplify the mooring setup. The final three lines
# display the mooring in 0.5m/s current.
# Exercise: add more floation between instruments,
# to stiffen this floppy mooring.
A <- anchor(depth=100)
W <- function(l) wire("3/8in wire/jack", length=l)
I <- instrument("SBE37 microcat clamp-on style")
F <- float("HMB 20")
m <- mooring(A, W(20), I, W(20), I, W(20), I, W(20), F)
md <- discretise(m)
mdk <- knockdown(md, 0.5)
plot(mdk)

}
\author{
Dan Kelley
}
