% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{mooring}
\alias{mooring}
\title{Create a mooring}
\usage{
mooring(...)
}
\arguments{
\item{...}{two or more elementary objects, as created by
\code{\link[=anchor]{anchor()}}, \code{\link[=release]{release()}}, \code{\link[=chain]{chain()}},
\code{\link[=wire]{wire()}}, \code{\link[=connector]{connector()}}, \code{\link[=instrument]{instrument()}}, \code{\link[=misc]{misc()}} or \code{\link[=float]{float()}}.}
}
\description{
Assemble elementary components into a mooring string.
The first argument must be an anchor, created
with \code{\link[=anchor]{anchor()}}.  These are stored in reverse order
in the return value (see Example 1).
}
\examples{
# Example 1: most basic case: anchor, line, float.
library(mooring)
m <- mooring(anchor(depth = 100), wire(length = 80), float())
m # whole-mooring overview
head(m, 1) # float overview

# Example 2: compare a simple mooring (with inadequate
# buoyancy) with a stiffer one that has extra buoyancy
# near an instrument.  Note that the stiffness is
# concentrated in the depth region that is being sampled
# with the instrument (a "microcat").
A <- anchor(depth = 150)
W <- function(l) wire("3/8in wire/jack", length = l)
MC <- instrument("SBE37 microcat clamp-on style")
BUB <- float("BUB 2x17in glass")
top <- float("Trpl 16in Viny")
m1 <- mooring(A, W(45), W(5), MC, W(5), W(80), top)
m2 <- mooring(A, W(45), BUB, W(5), MC, W(5), BUB, W(80), top)
m1k <- m1 |>
    discretise() |>
    knockdown(u = 0.5)
m2k <- m2 |>
    discretise() |>
    knockdown(u = 0.5)
plot(m1k)
# Show stiffened mooring in red
lines(x(m2k), depth(m2k), col = 2, lwd = 2)
points(x(m2k, skipWire = TRUE), depth(m2k, skipWire = TRUE), col = 2, pch = 20)
type <- sapply(m2k, \(i) class(i)[2])[!is.wire(m2k)]
typeAbbrev <- unname(sapply(type, \(t)
switch(t,
    float = "F",
    instrument = "I",
    anchor = "A"
)))
text(x(m2k, skipWire = TRUE), depth(m2k, skipWire = TRUE), typeAbbrev, col = 2, pos = 2)

}
\author{
Dan Kelley
}
