% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\docType{package}
\name{mooring}
\alias{mooring}
\title{mooring: A Package for Analysing Oceanographic Moorings.}
\description{
The mooring package provides functions for working with
oceanographic moorings.
}
\examples{
# Create, summarize, and plot a simple mooring comprising
# a bottom anchor, a 100-metre cable, and a float.
library(mooring)
A <- anchor()
C <- cable(100)
F <- float()
m <- A + C + F
print(m)
plot(m)

}
