% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{mooring}
\alias{mooring}
\title{Create a mooring}
\usage{
mooring(..., type = "bottom")
}
\arguments{
\item{...}{two or more elementary objects, as created by
\code{\link[=anchor]{anchor()}}, \code{\link[=release]{release()}}, \code{\link[=chain]{chain()}},
\code{\link[=wire]{wire()}}, \code{\link[=connector]{connector()}}, \code{\link[=instrument]{instrument()}}, \code{\link[=misc]{misc()}} or \code{\link[=float]{float()}}.
FIXME: discuss order here (and in relation to 'surface' argument).}

\item{type}{character value indicating the type of mooring.
The default, \code{"bottom"}, indicates a mooring attached to
the bottom.  The other alternative, \code{"surface"}, is not handled
yet, but the plan is to handle moorings that are attached to
the surface, via a float, attachment to ice, etc.}
}
\description{
Assemble elementary components into a mooring string.
The first argument must be an anchor, created
with \code{\link[=anchor]{anchor()}}.  These are stored in reverse order
in the return value (see Example 1).
}
\examples{
# Example 1: most basic case: anchor, line, float.
library(mooring)
m <- mooring(anchor(depth=100), wire(length=80), float())
m                          # overview of float
tail(m, 1)                 # anchor overview
head(m, 1)                 # float overview
str(m[1])                  # float details

# Example 2: mooring with mid-depth instrument and a float
# just below it.  Note the definition of some functions,
# to simplify the mooring setup. The final three lines
# display the mooring in 0.5m/s current.
# Exercise: add more floation between instruments,
# to stiffen this floppy mooring.
A <- anchor(depth=100)
W <- function(l) wire("3/8in wire/jack", length=l)
I <- instrument("SBE37 microcat clamp-on style")
F <- float("HMB 20")
m <- mooring(A, W(20), I, W(20), I, W(20), I, W(20), F)
md <- discretise(m)
mdk <- knockdown(md, 0.5)
plot(mdk)

}
\author{
Dan Kelley
}
