% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\docType{data}
\name{mooringElements}
\alias{mooringElements}
\title{Dataset of wire, chain, and float properties}
\usage{
data(mooringElements)
}
\description{
This dataset was constructed from the \code{mdcodes.mat} (Dewey 1999, 2021).
This is provided for use by \code{\link[=wire]{wire()}}, \code{\link[=chain]{chain()}} and \code{\link[=float]{float()}}.
Users will seldom need to consult this table directly, but those that
do should note that Dewey's lengths in cm have been converted to m.
}
\examples{
library(mooring)
data(mooringElements)
mooringElements

}
\references{
Dewey, Richard K. "Mooring Design & Dynamics-a Matlab® Package for
Designing and Analyzing Oceanographic Moorings." Marine Models, vol. 1, no. 1
(December 1, 1999): 103–57. https://doi.org/10.1016/S1369-9350(00)00002-X

Dewey, Richard. "Mooring Design and Dynamics:
A Matlab Package for Designing and Testing
Oceanographic Moorings And Towed Bodies."
Accessed May 15, 2021.
http://canuck.seos.uvic.ca/rkd/mooring/mdd/mdd.php
http://canuck.seos.uvic.ca/rkd/mooring/moordyn.php
}
