% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{plot.mooring}
\alias{plot.mooring}
\title{Plot a mooring object}
\usage{
\method{plot}{mooring}(x, which = "shape", showDepths = TRUE, fancy = FALSE, title = "", ...)
}
\arguments{
\item{x}{an object of the \code{"mooring"} class.}

\item{which}{character value indicating the desired plot, with
choices: \code{"shape"} and \code{"tension"}.}

\item{showDepths}{logical value indicating whether to indicate the depths of
floats, to the left of the symbols.}

\item{fancy}{logical value indicating whether to indicate the
water and sediments with filled rectangles.  The alternative
is a simpler plot.}

\item{title}{character value indicating a title to put above
the plot.}

\item{...}{optional arguments.}
}
\description{
Plot a mooring object
}
\examples{
# Create, summarize, and plot a simple mooring comprising
# a bottom anchor, a 100-metre wire, and a float.
library(mooring)
m <- mooring(anchor(depth=100), wire(length=80), float("HMB 20"))
md <- discretise(m)
par(mfrow=c(1, 2))
plot(md)
plot(md, which="tension")

}
\author{
Dan Kelley
}
