% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mooring}
\alias{plot.mooring}
\title{Plot a mooring}
\usage{
\method{plot}{mooring}(
  x,
  which = "shape",
  showInterfaces = TRUE,
  showDepths = FALSE,
  showLabels = TRUE,
  showDetails = FALSE,
  fancy = FALSE,
  title = "",
  mar = c(1.5, 3.5, 3.5, 1),
  mgp = c(2, 0.7, 0),
  xlim = NULL,
  type = "l",
  ...
)
}
\arguments{
\item{x}{an object of the \code{"mooring"} class.}

\item{which}{character value indicating the desired plot, with
choices: \code{"shape"} (the default), \code{"knockdown"}, \code{"tension"} and \code{"velocity"}.}

\item{showInterfaces}{logical value indicating whether to indicate the water
surface with a blue line and the ocean bottom with a brown line.}

\item{showDepths}{logical value indicating whether to indicate the depths of
floats, to the left of the symbols.}

\item{showLabels}{logical value indicating whether to indicate anchors,
instruments and floats with \code{A}, \code{I} and \code{F}, respectively.}

\item{showDetails}{logical value indicating whether to show details for anchors,
instruments and floats with text.}

\item{fancy}{logical value indicating whether to indicate the
water and sediments with filled rectangles.  The alternative
is a simpler plot.}

\item{title}{character value indicating a title to put above
the plot.}

\item{mar}{numeric vector of length 4, used to set margins.
The default value makes allowance for the axis along the top,
and tightens the margins on bottom and right.}

\item{mgp}{numeric vector of length 3, used to set axis
geometry.  The default value moves numbers and labels closer
to the axes than the usual R default.}

\item{xlim}{optional numeric vector of length 2 that can
be used to specify the limits of the horizontal axis.}

\item{type}{character value indicating type of plot. The default, \code{"l"},
means to draw lines, while e.g. \code{"p"} means to draw points.}

\item{...}{optional arguments.}
}
\description{
Plot a mooring
}
\examples{
# Create, summarize, and plot a simple mooring comprising
# a bottom anchor, a 100-metre wire, and a float.
library(mooring)
m <- mooring(anchor(depth = 100), wire(length = 80), float("HMB 20"))
md <- discretise(m)
par(mfrow = c(1, 2))
plot(md)
plot(md, which = "tension")

}
\author{
Dan Kelley
}
