% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{plot.mooring}
\alias{plot.mooring}
\title{Plot a mooring object}
\usage{
\method{plot}{mooring}(x, ...)
}
\arguments{
\item{x}{an object of the \code{"mooring"} class.}

\item{...}{optional arguments.}
}
\description{
Plot a mooring object
}
\examples{
# Create, summarize, and plot a simple mooring comprising
# a bottom anchor, a 100-metre wire, and a float.
library(mooring)
m <- anchor(depth=100) + wire(length=80) + float("HMB 20")
plot(m)

}
\author{
Dan Kelley
}
