% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{plot.mooring}
\alias{plot.mooring}
\title{Plot a mooring object}
\usage{
\method{plot}{mooring}(
  x,
  which = "shape",
  showDepths = TRUE,
  showLabels = TRUE,
  fancy = FALSE,
  title = "",
  mar = c(1.5, 3.5, 3.5, 1),
  mgp = c(2, 0.7, 0),
  ...
)
}
\arguments{
\item{x}{an object of the \code{"mooring"} class.}

\item{which}{character value indicating the desired plot, with
choices: \code{"shape"} and \code{"tension"}.}

\item{showDepths}{logical value indicating whether to indicate the depths of
floats, to the left of the symbols.}

\item{showLabels}{logical value indicating whether to indicate anchors,
instruments and floats with \code{A}, \code{I} and \code{F}, respectively.}

\item{fancy}{logical value indicating whether to indicate the
water and sediments with filled rectangles.  The alternative
is a simpler plot.}

\item{title}{character value indicating a title to put above
the plot.}

\item{mar}{numeric vector of length 4, used to set margins.
The default value makes allowance for the axis along the top,
and tightens the margins on bottom and right.}

\item{mgp}{numeric vector of length 3, used to set axis
geometry.  The default value moves numbers and labels closer
to the axes than the usual R default.}

\item{...}{optional arguments.}
}
\description{
Plot a mooring object
}
\examples{
# Create, summarize, and plot a simple mooring comprising
# a bottom anchor, a 100-metre wire, and a float.
library(mooring)
m <- mooring(anchor(depth=100), wire(length=80), float("HMB 20"))
md <- discretise(m)
par(mfrow=c(1, 2))
plot(md)
plot(md, which="tension")

}
\author{
Dan Kelley
}
