% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{+.mooring}
\alias{+.mooring}
\title{Combine two mooring objects}
\usage{
\method{+}{mooring}(m1, m2)
}
\arguments{
\item{m1, m2}{objects of the \code{"mooring"} class.}
}
\description{
The first object is "placed" the second.
}
\author{
Dan Kelley
}
