% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{+.mooring}
\alias{+.mooring}
\title{Combine two mooring objects (DEFUNCT)}
\usage{
\method{+}{mooring}(m1, m2)
}
\arguments{
\item{m1, m2}{DEFUNCT.}
}
\value{
DEFUNCT
}
\description{
This has been replaced by \code{\link[=mooring]{mooring()}}.
}
\author{
Dan Kelley
}
