% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{print.mooring}
\alias{print.mooring}
\title{Print a mooring object}
\usage{
\method{print}{mooring}(m)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class.}
}
\description{
Print a mooring object
}
\author{
Dan Kelley
}
