% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{print.mooring}
\alias{print.mooring}
\title{Print a mooring object}
\usage{
\method{print}{mooring}(x, ...)
}
\arguments{
\item{x}{an object of the \code{"mooring"} class.}

\item{...}{optional arguments.}
}
\description{
Print a mooring object
}
\examples{
library(mooring)
m <- anchor() + wire(height=100) + float()
print(m)
}
\author{
Dan Kelley
}
