% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release}
\alias{release}
\title{Create a mooring-release object}
\usage{
release(
  model = "EG&G 723a",
  buoyancy = NULL,
  height = NULL,
  area = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the type of release.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If this is \code{"?X"}, where X is a set of
characters, then \code{\link[=findElement]{findElement()}} is called to do a fuzzy search,
with the \code{search} argument set to \code{"release"}.
(3) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(4) Otherwise, a new release object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancy}{numeric value indicating buoyancy.  For
wire elements, this is buoyancy per length, expressed in kg/m.
For other elements, it is buoyancy, expressed in kg.
Note that \code{buoyancy} is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height (in m) of the element.
This value is used in the computation of the mooring geometry.
In some cases, as e.g. in Dewey-derived data that have "diameter" tabulated
as zero, it is also used in the computation of the area
of the element, as height*width.
Note that \code{height} is ignored if \code{model} is recognized.}

\item{area}{a numerical value indicating the horizontal instrument area
in \eqn{m^2}{m^2}.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}
}
\value{
\code{release} returns an object of the \code{"mooringElement"} class and \code{"release"} subclass.
}
\description{
Create a mooring-release object,
either by looking up a known object from the database, or by defining a new type.
Area is formulated as length*width, for consistency with Dewey's Matlab code.
}
\section{Data sources}{


To learn more about the data used for the built-in
release
objects, use e.g.
\preformatted{
data(mooringElements)
mooringElements$releases
}
}

\examples{
library(mooring)
# List known wire types
release("?")

}
\seealso{
Other functions that create mooring elements: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{connector}()},
\code{\link{float}()},
\code{\link{instrument}()},
\code{\link{misc}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring elements}
