% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{release}
\alias{release}
\title{Create a mooring-release object}
\usage{
release(
  model = "eg&g 723a",
  buoyancy = NULL,
  height = NULL,
  width = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the type of release.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If this is \code{"?X"}, where X is a set of
characters, then \code{\link[=findElement]{findElement()}} is called to do a fuzzy search,
with the \code{search} argument set to \code{"release"}.
(3) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(4) Otherwise, a new release object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancy}{numeric value indicating buoyancy.  For
wire elements, this is buoyancy per length, expressed in kg/m.
For other elements, it is buoyancy, expressed in kg.
Note that \code{buoyancy} is ignored if \code{model} is recognized.}

\item{height}{numeric value for the height (in m) of the element.
This value is used in the computation of (a) the mooring geometry
and (b) the area of the element (which is in turn used in
the drag computation).
Note that \code{height} is ignored if \code{model} is recognized.}

\item{width}{numeric value for the width (in m)
of the element as viewed from above.
This is multiplied by \code{height} to compute the
area of the element, as part of the drag computation.
Note that \code{width} is ignored if \code{model} is recognized.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}
}
\value{
\code{release} returns an object of the \code{"mooring"} class and \code{"release"} subclass.
}
\description{
Create a mooring-release object,
either by looking up a known object from the database, or defining a new type.
}
\examples{
library(mooring)
# List known wire types
release("?")

}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{connector}()},
\code{\link{float}()},
\code{\link{instrument}()},
\code{\link{wire}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
