% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\name{summary}
\alias{summary}
\title{Summarize a mooring or a mooringElement}
\arguments{
\item{x}{either a \code{mooring} object, created with \code{\link[=mooring]{mooring()}},
or a \code{mooringElement} object, created with \code{\link[=anchor]{anchor()}}, \code{\link[=wire]{wire()}},
or any of the related functions that create mooring elements.}
}
\description{
Summarize a mooring or a mooringElement

Summarize a mooring element
}
\examples{
library(mooring)
a <- anchor()
summary(a)

m <- mooring(anchor(), wire(length = 80), float("HMB 20"), waterDepth = 100)
summary(m)

}
\author{
Dan Kelley
part of the item.

Dan Kelley
}
