% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{tension}
\alias{tension}
\title{Return tension, in kg, between elements in mooring.}
\usage{
tension(m, stagnant = FALSE)
}
\arguments{
\item{m}{an object of the \code{"mooring"} class.}

\item{stagnant}{a logical indicating whether to compute tension
as though no currents were present.}
}
\value{
a numeric vector of tension, in kg.
}
\description{
The first element (for the anchor) is repeated, so that the
length of the returned result matches the length of \code{m}.
}
\examples{
library(mooring)
m <- anchor(depth=120) + wire(length=100) + float("HMB 20")
md <- discretise(m)
mdk <- knockdown(md, u=0.5)
depth <- depth(mdk)
# Next is handled better by plot.mooring(mdk,which="tension")
plot(tension(mdk), depth, ylim=rev(range(depth)), type="l")

}
\author{
Dan Kelley
}
