% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{wire}
\alias{wire}
\title{Create a wire object}
\usage{
wire(
  model = "1/4 wire/jack",
  buoyancyPerMeter = NULL,
  diameter = NULL,
  CD = NULL,
  length = NULL
)
}
\arguments{
\item{model}{character value indicating the type of wire.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(3) Otherwise, a new wire object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancyPerMeter}{numeric value indicating buoyancy per meter
of length, expressed as kg/m.
Note that \code{buoyancy} is ignored if \code{model} is recognized.}

\item{diameter}{numeric value for the diameter (in m)
of the element as viewed from above.
This is multiplied by \code{height} to compute the
area of the element, as part of the drag computation.
Note that \code{diameter} is ignored if \code{model} is recognized.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}

\item{length}{numeric value indicating the length (in m) of the wire.}
}
\value{
\code{wire} returns an object of the \code{"mooring"} class and \code{"wire"} subclass.
}
\description{
Creates an object that describes a mooring wire, or other type of line, e.g.
nylon ropes are considered wires, for this purpose.
Note that all dimensions are specified in m, not cm.
}
\examples{
library(mooring)
# List known wire types
wire("?")

}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{float}()},
\code{\link{instrument}()},
\code{\link{release}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
