% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{wire}
\alias{wire}
\title{Create a wire object}
\usage{
wire(
  model = "1/4 wire/jack",
  buoyancy = NULL,
  length = NULL,
  width = NULL,
  CD = NULL
)
}
\arguments{
\item{model}{character value indicating the model name of the wire. The
default, \code{"1/4 wire/jack"}, is for quarter-inch jacketed steel wire.
Use \code{wire("?")} to get a listing of permitted values.
If \code{model} is not recognized, then
\link{wire} creates a new type, using the stated diameter and CD values;
otherwise, these arguments are ignored.}

\item{buoyancy}{numeric value of the buoyancy per length of wire, in kg/m.}

\item{length}{numeric value indicating the length of the wire, in metres.}

\item{width}{numeric value for the width of the wire, in m.
This is ignored if \code{model} is recognized.}

\item{CD}{numeric value for the drag coefficient for the wire.
This is ignored if \code{model} is recognized.}
}
\value{
\link{wire} returns an object of the \code{"mooring"} class, with \code{type} equal to \code{"wire"}.
}
\description{
Creates an object that describes a mooring wire, or other type of line, e.g.
nylon ropes are considered wires, for this purpose.
Note that all dimensions are specified in m, not cm.
}
\examples{
library(mooring)
# List known wire types
wire("?")

}
\seealso{
Other functions that create mooring objects: 
\code{\link{+.mooring}()},
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{float}()},
\code{\link{release}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
