% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{wire}
\alias{wire}
\title{Create a wire object}
\usage{
wire(
  model = "1/4in wire/jack",
  buoyancyPerMeter = NULL,
  areaPerMeter = NULL,
  CD = NULL,
  length = NULL
)
}
\arguments{
\item{model}{character value indicating the type of wire.
There are three possibilities for \code{model}.
(1) If this is \code{"?"}, then the function returns a vector of permitted
character values. (2) If this is \code{"?X"}, where X is a set of
characters, then \code{\link[=findElement]{findElement()}} is called to do a fuzzy search,
with the \code{search} argument set to \code{"wire"}.
(3) If it is recognized, i.e. if that type is
stored in \code{data(mooringElements)}, then the that stored value
is used, and all other arguments to this function are ignored.
(4) Otherwise, a new wire object is created, using
values specified in the other arguments (all of which must
be supplied).}

\item{buoyancyPerMeter}{numeric value indicating buoyancy per meter
of length, expressed as kg/m.
Note that \code{buoyancyPerMeter} is ignored if \code{model} is recognized.}

\item{areaPerMeter}{numeric value indicating the frontal area per meter
of length, expressed as m^2. (This is equivalent to the width, or diameter, of
the wire.)
Note that \code{areaPerMeter} is ignored if \code{model} is recognized.}

\item{CD}{numeric value (unitless) for the drag coefficient.
Note that \code{CD} is ignored if \code{model} is recognized.}

\item{length}{(mandatory) numeric value indicating the length (in m) of the wire.}
}
\value{
\code{wire} returns an object of the \code{"mooring"} class and \code{"wire"} subclass.
}
\description{
Creates a wire (or rope, chain, etc.) object,
either by looking up a known object from the database, or by defining a new type.
}
\section{Data sources}{


To learn more about the data used for the built-in
wire
objects, use e.g.
\preformatted{
data(mooringElements)
mooringElements$wires
}
}

\examples{
library(mooring)
# List known wire types
wire("?")

}
\seealso{
Other functions that create mooring objects: 
\code{\link{anchor}()},
\code{\link{chain}()},
\code{\link{connector}()},
\code{\link{float}()},
\code{\link{instrument}()},
\code{\link{misc}()},
\code{\link{release}()}
}
\author{
Dan Kelley
}
\concept{functions that create mooring objects}
