% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooring.R
\name{x}
\alias{x}
\title{Return horizontal coordinate, in m, of elements in mooring.}
\usage{
x(m, stagnant = FALSE)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{stagnant}{logical value indicating whether to return
values for the mooring configuration in the stagnant case
(with zero velocity) or the present case (possibly with
nonzero velocity).}
}
\value{
a numeric vector of horizontal coordinate in metres.
}
\description{
Return horizontal coordinate, in m, of elements in mooring.
}
\examples{
library(mooring)
m <- mooring(anchor(depth=120), wire(length=100), float("HMB 20"))
x(m)

}
\author{
Dan Kelley
}
