% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\name{z}
\alias{z}
\title{Vertical coordinate of mooring elements}
\usage{
z(m, stagnant = FALSE, skipWire = FALSE)
}
\arguments{
\item{m}{a mooring object, created by \code{\link[=mooring]{mooring()}}.}

\item{stagnant}{logical value indicating whether to return
values for the mooring configuration in the stagnant case
(with zero velocity) or the present case (possibly with
nonzero velocity).}

\item{skipWire}{logical value indicating whether to skip wire elements, i.e. to return values just for floats, instruments, etc.}
}
\value{
a numeric vector of vertical coordinate in metres.
}
\description{
This is the z coordinate of the \emph{top} of the element. See also
\code{\link[=depth]{depth()}}, which is the negative of the result from \code{z()}.
}
\examples{
library(mooring)
m <- mooring(anchor(depth = 120), wire(length = 100), float("HMB 20"))
z(m)

}
\author{
Dan Kelley
}
